renamed module generic_joint -> joint_candidate
