+++ src/compas_timber/connections/joint_candidate.py
